requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
cryptography==43.0.1
dataclasses-json==0.6.7
deep-translator==1.11.4
Deprecated==1.2.14
discord==2.3.2
discord.py==2.4.0
disnake==2.9.2
distro==1.9.0
duckduckgo_search==6.2.13
edge-tts==6.1.12
fastapi==0.115.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
groq==0.11.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-groq==0.2.0
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.129
lxml==5.3.0
lyrics-extractor==3.0.1
lyricsgenius==3.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
ollama==0.3.3
orjson==3.10.7
packaging==24.1
pillow==10.4.0
primp==0.6.3
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyGithub==2.4.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pypiwin32==223
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
python-dotenv==1.0.1
pyttsx3==2.98
pytube==15.0.0
pywin32==306
PyYAML==6.0.2
redis==5.1.0
regex==2024.9.11
requests==2.32.3
rich==13.8.1
shiboken6==6.7.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spotipy==2.24.0
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==8.5.0
timeout==0.1.2
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
websocket==0.2.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.13.1
youtube-transcript-api==0.6.2
yt-dlp==2024.9.27
zope.event==5.0
zope.interface==7.0.3
